Virtual Page Allocator

+++ Nominax/Source/Foundation/OSIImpl_Windows.cpp
@@ -350,7 +350,7 @@ namespace Nominax::Foundation
 		return result;
 	}
 
-	auto OSI::MemoryUnmap(void* const region, [[maybe_unused]] const std::size_t size) -> bool
+	auto OSI::MemoryUnmap(void* const region, [[maybe_unused]] const std::uint64_t size) -> bool
 	{
 		const LPVOID argAddress {region};
 		constexpr DWORD argFreeType {MEM_RELEASE};
